+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel_monitor.rs
@@ -8,6 +8,7 @@ use nym_sdk::UserAgent;
 use nym_task::TaskManager;
 use nym_vpn_account_controller::AccountStateReceiver;
 use nym_vpn_network_config::start_background_file_refresh;
+use nym_wg_metadata_client::TunUpSendData;
 #[cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))]
 use std::net::{Ipv4Addr, Ipv6Addr};
 #[cfg(any(target_os = "linux", target_os = "ios", target_os = "android"))]
@@ -65,7 +66,7 @@ use crate::tunnel_provider::OSTunProvider;
 use crate::tunnel_state_machine::route_handler::TUNNEL_FWMARK;
 use crate::{
     VpnTopologyProvider,
-    tunnel_state_machine::{WireguardMultihopMode, account, ipv6_availability},
+    tunnel_state_machine::{TunnelConstants, WireguardMultihopMode, account, ipv6_availability},
 };
 
 /// Default MTU for mixnet tun device.
@@ -186,6 +187,7 @@ pub struct TunnelParameters {
     pub nym_config: NymConfig,
     pub resolved_gateway_config: ResolvedConfig,
     pub tunnel_settings: TunnelSettings,
+    pub tunnel_constants: TunnelConstants,
     pub selected_gateways: Option<SelectedGateways>,
 }
 
@@ -716,7 +718,7 @@ impl TunnelMonitor {
         task_manager: &TaskManager,
         connected_mixnet: ConnectedMixnet,
     ) -> Result<StartTunnelResult> {
-        let connected_tunnel = connected_mixnet
+        let (connected_tunnel, tun_up) = connected_mixnet
             .connect_wireguard_tunnel(
                 task_manager,
                 &self.tunnel_parameters.nym_config.network_env,
@@ -751,9 +753,15 @@ impl TunnelMonitor {
         });
 
         let dns_config = self.tunnel_parameters.tunnel_settings.resolved_dns_config();
+        let initial_allowed_ips = self
+            .tunnel_parameters
+            .tunnel_constants
+            .initial_allowed_ips
+            .clone();
         let tunnel_options = TunnelOptions::Netstack(NetstackTunnelOptions {
             exit_tun,
             dns: dns_config.tunnel_config().to_vec(),
+            initial_allowed_ips,
         });
 
         let tunnel_metadata = TunnelMetadata {
@@ -766,12 +774,14 @@ impl TunnelMonitor {
             ipv6_gateway: Some(conn_data.entry.private_ipv6),
         };
 
-        let tunnel_handle = AnyTunnelHandle::from(
-            connected_tunnel
-                .run(tunnel_options)
-                .await
-                .map_err(Box::new)?,
-        );
+        let tunnel_handle = connected_tunnel
+            .run(tunnel_options)
+            .await
+            .map_err(Box::new)?;
+        let tunnel_handle = AnyTunnelHandle::from(tunnel_handle);
+
+        let _ = tun_up.entry.send(TunUpSendData::Signal);
+        let _ = tun_up.exit.send(TunUpSendData::Signal);
 
         Ok(StartTunnelResult {
             tunnel_interface: TunnelInterface::One(tunnel_metadata),
@@ -786,7 +796,7 @@ impl TunnelMonitor {
         task_manager: &TaskManager,
         connected_mixnet: ConnectedMixnet,
     ) -> Result<StartTunnelResult> {
-        let connected_tunnel = connected_mixnet
+        let (connected_tunnel, tun_up) = connected_mixnet
             .connect_wireguard_tunnel(
                 task_manager,
                 &self.tunnel_parameters.nym_config.network_env,
@@ -828,11 +838,17 @@ impl TunnelMonitor {
         });
 
         let dns_config = self.tunnel_parameters.tunnel_settings.resolved_dns_config();
+        let initial_allowed_ips = self
+            .tunnel_parameters
+            .tunnel_constants
+            .initial_allowed_ips
+            .clone();
         let tunnel_options = TunnelOptions::Netstack(NetstackTunnelOptions {
             exit_tun_name: WG_EXIT_WINTUN_NAME.to_owned(),
             exit_tun_guid: WG_EXIT_WINTUN_GUID.to_owned(),
             wintun_tunnel_type: WINTUN_TUNNEL_TYPE.to_owned(),
             dns: dns_config.tunnel_config().to_vec(),
+            initial_allowed_ips,
         });
 
         let mut tunnel_handle = connected_tunnel
@@ -872,6 +888,9 @@ impl TunnelMonitor {
         // Update interface name in tunnel metadata
         tunnel_metadata.interface = wintun_exit_interface.name.clone();
 
+        let _ = tun_up.entry.send(TunUpSendData::Signal);
+        let _ = tun_up.exit.send(TunUpSendData::Signal);
+
         Ok(StartTunnelResult {
             tunnel_interface: TunnelInterface::One(tunnel_metadata),
             tunnel_handle: AnyTunnelHandle::from(tunnel_handle),
@@ -885,7 +904,7 @@ impl TunnelMonitor {
         task_manager: &TaskManager,
         connected_mixnet: ConnectedMixnet,
     ) -> Result<StartTunnelResult> {
-        let connected_tunnel = connected_mixnet
+        let (connected_tunnel, tun_up) = connected_mixnet
             .connect_wireguard_tunnel(
                 task_manager,
                 &self.tunnel_parameters.nym_config.network_env,
@@ -913,7 +932,7 @@ impl TunnelMonitor {
             ips.push(IpAddr::V6(conn_data.entry.private_ipv6));
         }
         let entry_tunnel_metadata = TunnelMetadata {
-            interface: entry_tun_name,
+            interface: entry_tun_name.clone(),
             ips,
             ipv4_gateway: None,
             ipv6_gateway: None,
@@ -963,18 +982,34 @@ impl TunnelMonitor {
         });
 
         let dns_config = self.tunnel_parameters.tunnel_settings.resolved_dns_config();
+        let initial_allowed_ips = self
+            .tunnel_parameters
+            .tunnel_constants
+            .initial_allowed_ips
+            .clone();
         let tunnel_options = TunnelOptions::TunTun(TunTunTunnelOptions {
             entry_tun,
             exit_tun,
             dns: dns_config.tunnel_config().to_vec(),
+            initial_allowed_ips,
         });
 
-        let tunnel_handle = AnyTunnelHandle::from(
-            connected_tunnel
-                .run(tunnel_options)
-                .await
-                .map_err(Box::new)?,
-        );
+        let tunnel_handle = connected_tunnel
+            .run(tunnel_options)
+            .await
+            .map_err(Box::new)?;
+        let tunnel_handle = AnyTunnelHandle::from(tunnel_handle);
+
+        let (entry_data, exit_data) = if cfg!(target_os = "linux") {
+            (
+                TunUpSendData::InterfaceName(entry_tun_name),
+                TunUpSendData::InterfaceName(exit_tun_name),
+            )
+        } else {
+            (TunUpSendData::Signal, TunUpSendData::Signal)
+        };
+        let _ = tun_up.entry.send(entry_data);
+        let _ = tun_up.exit.send(exit_data);
 
         Ok(StartTunnelResult {
             tunnel_interface: TunnelInterface::Two {
@@ -992,7 +1027,7 @@ impl TunnelMonitor {
         task_manager: &TaskManager,
         connected_mixnet: ConnectedMixnet,
     ) -> Result<StartTunnelResult> {
-        let connected_tunnel = connected_mixnet
+        let (connected_tunnel, tun_up) = connected_mixnet
             .connect_wireguard_tunnel(
                 task_manager,
                 &self.tunnel_parameters.nym_config.network_env,
@@ -1000,6 +1035,7 @@ impl TunnelMonitor {
             )
             .await
             .map_err(Box::new)?;
+
         let conn_data = connected_tunnel.connection_data();
         let entry_tun_mtu = connected_tunnel.entry_mtu();
         let exit_tun_mtu = connected_tunnel.exit_mtu();
@@ -1048,6 +1084,11 @@ impl TunnelMonitor {
         });
 
         let dns_config = self.tunnel_parameters.tunnel_settings.resolved_dns_config();
+        let initial_allowed_ips = self
+            .tunnel_parameters
+            .tunnel_constants
+            .initial_allowed_ips
+            .clone();
         let tunnel_options = TunnelOptions::TunTun(TunTunTunnelOptions {
             entry_tun_name: WG_ENTRY_WINTUN_NAME.to_owned(),
             entry_tun_guid: WG_ENTRY_WINTUN_GUID.to_owned(),
@@ -1055,6 +1096,7 @@ impl TunnelMonitor {
             exit_tun_guid: WG_EXIT_WINTUN_GUID.to_owned(),
             wintun_tunnel_type: WINTUN_TUNNEL_TYPE.to_owned(),
             dns: dns_config.tunnel_config().to_vec(),
+            initial_allowed_ips,
         });
 
         let mut tunnel_handle = connected_tunnel
@@ -1117,6 +1159,9 @@ impl TunnelMonitor {
             return Err(err);
         }
 
+        let _ = tun_up.entry.send(TunUpSendData::Signal);
+        let _ = tun_up.exit.send(TunUpSendData::Signal);
+
         Ok(StartTunnelResult {
             tunnel_interface,
             tunnel_handle: AnyTunnelHandle::from(tunnel_handle),
@@ -1130,7 +1175,7 @@ impl TunnelMonitor {
         task_manager: &TaskManager,
         connected_mixnet: ConnectedMixnet,
     ) -> Result<StartTunnelResult> {
-        let connected_tunnel = connected_mixnet
+        let (connected_tunnel, tun_up) = connected_mixnet
             .connect_wireguard_tunnel(
                 task_manager,
                 &self.tunnel_parameters.nym_config.network_env,
@@ -1189,16 +1234,25 @@ impl TunnelMonitor {
             .dns
             .ip_addresses(&self.tunnel_parameters.tunnel_settings.default_dns_ips())
             .to_vec();
+        let initial_allowed_ips = self
+            .tunnel_parameters
+            .tunnel_constants
+            .initial_allowed_ips
+            .clone();
 
         let tunnel_handle = connected_tunnel
             .run(
                 tun_device,
                 dns_servers,
+                initial_allowed_ips,
                 #[cfg(target_os = "android")]
                 self.tun_provider.clone(),
             )
             .await
             .map_err(Box::new)?;
+
+        let _ = tun_up.entry.send(TunUpSendData::Signal);
+        let _ = tun_up.exit.send(TunUpSendData::Signal);
 
         Ok(StartTunnelResult {
             tunnel_conn_data,